regenerate-static-content: Allow decoration that's not symmetric in Y
Inkscape has a mismatch between the coordinates reported by `--query`
and those expected from `--export-area`. For the former, the origin is
in the top-left corner. For the latter, it is in the bottom-right
corner.

As decorations were centered in Y-direction up to now, this mismatch
did not come into play up to now. But in upcoming commits, we'll shift
the decoration downwards a bit to allow for holes from hole punches.
So we need to set the Y-parameter straight that we pass to
`--export-area`.