Scanner: Take mask.offset into account for cropping
mask.offset was used when generating the mask. But during cropping, we
only focused on pixel values. This meant that the extracted image and
coloring page went out of sync a bit. This showed especially for thin
details.

We now compute the crop borders on the image before applying
mask.offset. And we compute it before-hand to save a bit of runtime
when scanning. This should give both better and more reliable results
in shorter time.