tests: Be less strict when asserting colors
Asserting a color without any deviation sometimes fails due to
artifacts (E.g.: jpg) when generating the files. So we switch to an
allowed deviation of 0.01, which allows an effective deviation of ~3
in the 0-255 scale. This allows to match a desired 255 with an actual
value of 254, which we saw across different OSes.